// src/api/sqlmodule/model/sqlQueryResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

/**
 * Ответ с данными эталонного SQL-запроса.
 */
export interface SqlQueryResponse {
  /** Уникальный идентификатор запроса. */
  id?: string;
  /** Идентификатор целевой БД (датасет). */
  targetDbId?: string;
  /** Текст SQL-запроса. */
  queryText?: string;
  /** Признак обязательного порядка колонок. */
  strictColumnOrder?: boolean;
  /** Признак обязательного порядка строк. */
  strictRowOrder?: boolean;
  /** Колонки золотого результата. */
  expectedColumns?: string[];
  /** Строки золотого результата. */
  expectedRows?: string[][];
  /** Идентификатор пользователя, создавшего запись. */
  createdById?: string;
  /** Дата и время создания записи. */
  createdAt?: string;
  /** Идентификатор пользователя, последним изменившего запись. */
  updatedById?: string;
  /** Дата и время последнего изменения записи. */
  updatedAt?: string;
}